+++ src/components/partials/value.js
@@ -1,39 +1,40 @@
 import css from 'dom-css';
 
 export default (root, text, theme, width, left) => {
-    let background = root.appendChild(document.createElement('div'));
-    let value = background.appendChild(document.createElement('span'));
 
-    value.innerHTML = text;
+    let input = root.appendChild(document.createElement('input'))
+    input.type = 'text'
 
-    let bgcss = {
+    input.value = text;
+
+    let inputCss = {
         position: 'absolute',
         backgroundColor: theme.colors.componentBackground,
-        paddingLeft: '1.5%',
+        paddingLeft: '1%',
         height: '20px',
         width: width,
         display: 'inline-block',
-        overflow: 'hidden'
-    }
+        overflow: 'hidden',
+        border: 'none',
 
-    if (!left) {
-        bgcss.right = 0;
-    }
-
-    css(background, bgcss);
-
-    css(value, {
         'font-family': `'Hack', monospace`,
         'font-size': '11px',
         color: theme.colors.textSecondary,
-        display: 'inline-block',
         userSelect: 'text',
         cursor: 'text',
-        overflow: 'hidden',
         lineHeight: '20px',
         wordBreak: 'break-all',
-        height: 20
-    });
 
-    return value;
+        'box-sizing': 'border-box',
+        '-moz-box-sizing': 'border-box',
+        '-webkit-box-sizing': 'border-box'
+    }
+
+    if (!left) {
+        inputCss.right = 0;
+    }
+
+    css(input, inputCss);
+
+    return input;
 }